refactor payment request details section & create separate type component
