+++ src/components/paymentRequestDetails/PaymentRequestCategoryProperties.tsx
@@ -1,27 +1,8 @@
 import React from "react";
-import {
-  FormControl,
-  InputAdornment,
-  MenuItem,
-  Select,
-  SelectChangeEvent,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableRow,
-  TextField,
-} from "@mui/material";
-import { NoteInfo } from "../../pages/workspaceDashboard/newPaymentRequest/newPaymentRequest.style";
-import { Image } from "../../pages/workspace/paymentRequest/paymentRequest.style";
-import ReactSelect from "../ReactSelect";
-import usePaymentsStore from "../../store/usePayments";
-import selectIcon from "../../assets/workspace/select.svg";
-import multiSelect from "../../assets/workspace/multi-select.svg";
-import optionsIcon from "../../assets/workspace/option.svg";
-import { ReactSelectOption } from "../../pages/workspace/paymentRequest/PaymentRequestDetailsReadOnly";
 import SingleSelectType from "./SingleSelectType";
 import MultiSelectType from "./MultiSelectType";
+import TextType from "./TextType";
+import { ReactSelectOption } from "../../pages/workspace/paymentRequest/PaymentRequestDetails";
 
 interface PaymentRequestCategoryPropertiesProps {
   selectedCategory: any;
@@ -64,53 +45,6 @@ const PaymentRequestCategoryProperties = ({
     <>
       {selectedCategory?.properties?.map((property: any) => (
         <React.Fragment key={property.ID}>
-          {/* {property.type === "single-select" && !property.archived && (
-            <TableRow
-              sx={{
-                td: {
-                  border: "1px solid var(--border-table)",
-                  padding: 1,
-                  paddingInline: 1,
-                },
-              }}
-            >
-              <TableCell sx={{ height: 1, width: 200 }}>
-                <NoteInfo>
-                  <Image src={selectIcon} alt="" /> {property.name}
-                </NoteInfo>
-              </TableCell>
-              <TableCell onBlur={handleUpdateCategory}>
-                <ReactSelect
-                  isMulti={false}
-                  isDisabled={status === 2}
-                  value={selectSingleValue}
-                  onChange={(selectedOption: ReactSelectOption) =>
-                    handleSelectSingleChange(
-                      selectedOption,
-                      property.name,
-                      property.type
-                    )
-                  }
-                  options={property.values.split(";").map((v: string) => ({
-                    value: v,
-                    label: v,
-                  }))}
-                  defaultValues={parseCategoryProperties
-                    .filter(
-                      (p: any) =>
-                        p.type === "single-select" && p.name === property.name
-                    )
-                    .map((p: any) =>
-                      p.values.split(";").map((v: string) => ({
-                        value: v,
-                        label: v,
-                      }))
-                    )
-                    .flat()}
-                />
-              </TableCell>
-            </TableRow>
-          )} */}
           {property.type === "single-select" &&
             (parseCategoryProperties.length > 0 ? (
               <SingleSelectType
@@ -165,80 +99,22 @@ const PaymentRequestCategoryProperties = ({
         <>
           {property.type === "Text" &&
             (parseCategoryProperties.length > 0 ? (
-              <TableRow
-                sx={{
-                  td: {
-                    border: "1px solid var(--border-table)",
-                    padding: 1,
-                    paddingInline: 1,
-                  },
-                }}
-              >
-                <TableCell sx={{ height: 1, width: 200 }}>
-                  <NoteInfo>
-                    <Image src={optionsIcon} alt="" /> {property.name}
-                  </NoteInfo>
-                </TableCell>
-
-                <TableCell onBlur={handleUpdateCategory}>
-                  <TextField
-                    disabled={status === 2}
-                    sx={{
-                      "& fieldset": { border: "none" },
-                    }}
-                    size="small"
-                    fullWidth
-                    // value={propertyContent}
-                    value={proPertyTextValue[property.name]?.values || ""}
-                    // id="fullWidth"
-                    placeholder="Enter content"
-                    onChange={(e) =>
-                      handlePropertyText(e, property.name, property.type)
-                    }
-                    InputProps={{
-                      style: { padding: 0 },
-                    }}
-                  />
-                </TableCell>
-              </TableRow>
+              <TextType
+                property={property}
+                handleUpdateCategory={handleUpdateCategory}
+                proPertyTextValue={proPertyTextValue}
+                handlePropertyText={handlePropertyText}
+                status={status}
+              />
             ) : (
               !property.archived && (
-                <TableRow
-                  sx={{
-                    td: {
-                      border: "1px solid var(--border-table)",
-                      padding: 1,
-                      paddingInline: 1,
-                    },
-                  }}
-                >
-                  <TableCell sx={{ height: 1, width: 200 }}>
-                    <NoteInfo>
-                      <Image src={optionsIcon} alt="" /> {property.name}
-                    </NoteInfo>
-                  </TableCell>
-
-                  <TableCell onBlur={handleUpdateCategory}>
-                    <TextField
-                      disabled={status === 2}
-                      sx={{
-                        "& fieldset": { border: "none" },
-                      }}
-                      size="small"
-                      fullWidth
-                      // value={propertyContent}
-                      value={proPertyTextValue[property.name]?.values || ""}
-                      // id="fullWidth"
-                      placeholder="Enter content"
-                      onChange={(e) =>
-                        handlePropertyText(e, property.name, property.type)
-                      }
-                      InputProps={{
-                        style: { padding: 0 },
-                      }}
-                    />
-                  </TableCell>
-                </TableRow>
+                <TextType
+                  property={property}
+                  handleUpdateCategory={handleUpdateCategory}
+                  proPertyTextValue={proPertyTextValue}
+                  handlePropertyText={handlePropertyText}
+                  status={status}
+                />
               )
             ))}
         </>
